viral-effect-app/tools/trend-collector/requirements.txt
requests>=2.31.0
feedparser>=6.0.11
python-dotenv>=1.0.0
